Test refactoring and support for "X" pattern (instead of "u")
