+++ .gitignore
@@ -6,3 +6,4 @@
 **/obj/
 /packages/*
 **/bin/
+/Edtf.sln.DotSettings.user
