+++ EdtfTests/L2MultipleDates.cs
@@ -1,0 +1,8 @@
+﻿using NUnit.Framework;
+
+namespace EdtfTests
+{
+    [TestFixture()] public class L2MultipleDates {
+        // TODO
+    }
+}
